+++ src/core/component/flyweight/index.ts
@@ -11,26 +11,26 @@
  * @packageDocumentation
  */
 
-import { defProp } from 'core/const/props';
 import { deprecate } from 'core/functools/deprecation';
-import { components, NULL } from 'core/component/const';
+
+import { components } from 'core/component/const';
+import { supports, CreateElement, VNode } from 'core/component/engines';
+
+import * as init from 'core/component/construct';
 
 import { initProps } from 'core/component/prop';
 import { initFields } from 'core/component/field';
+import { destroyComponent, FlyweightVNode } from 'core/component/functional';
 
-import { getNormalParent } from 'core/component/traverse';
-import { attachAccessorsFromMeta } from 'core/component/accessor';
 import { attachMethodsFromMeta } from 'core/component/method';
 import { implementEventAPI } from 'core/component/event';
+import { attachAccessorsFromMeta } from 'core/component/accessor';
 
-import { supports, CreateElement, VNode } from 'core/component/engines';
+import { getNormalParent } from 'core/component/traverse';
 import { getComponentDataFromVNode, patchComponentVData } from 'core/component/vnode';
 import { execRenderObject } from 'core/component/render';
 
 import { ComponentInterface } from 'core/component/interface';
-import { FlyweightVNode } from 'core/component/flyweight/interface';
-
-export * from 'core/component/flyweight/interface';
 
 /**
  * Takes a vnode and, if it has the composite attribute, returns a new vnode that contains a flyweight component,
@@ -72,32 +72,39 @@ export function parseVNodeAsFlyweight(
 		componentProto = meta.constructor.prototype,
 		componentTpl = TPLS[compositeAttr] ?? componentProto.render;
 
-	// To create a flyweight component we need to create a "fake" context for a component.
-	// The context is based on the specified parent context by using Object.create.
+	// To create a flyweight component we need to create the "fake" context.
+	// The context is based on the specified parent context by using `Object.create`.
 	// Also, we need to shim some component hooks.
 
 	const fakeCtx = Object.assign(Object.create(parentComponent), {
+		isFlyweight: true,
+
+		componentName: meta.componentName,
 		meta,
+
 		hook: 'beforeDataCreate',
 		instance: meta.instance,
-		componentName: meta.componentName,
-		isFlyweight: true
+
+		$destroy(): void {
+			destroyComponent(this);
+		}
 	});
 
 	fakeCtx.unsafe = fakeCtx;
+	fakeCtx.$createElement = createElement.bind(fakeCtx);
 
 	fakeCtx._self = fakeCtx;
 	fakeCtx._renderProxy = fakeCtx;
-
-	fakeCtx.$createElement = createElement.bind(fakeCtx);
 	fakeCtx._c = fakeCtx.$createElement;
 	fakeCtx._staticTrees = [];
 
-	attachMethodsFromMeta(fakeCtx);
-	implementEventAPI(fakeCtx);
-	attachAccessorsFromMeta(fakeCtx, true);
-
 	Object.defineProperty(fakeCtx, '$refs', {
+		configurable: true,
+		enumerable: true,
+		value: {}
+	});
+
+	Object.defineProperty(fakeCtx, '$refHandlers', {
 		configurable: true,
 		enumerable: true,
 		value: {}
@@ -158,63 +165,20 @@ export function parseVNodeAsFlyweight(
 		value: vnode.children
 	});
 
-	// Shim component input properties
-	for (let o = componentData.props, keys = Object.keys(o), i = 0; i < keys.length; i++) {
-		const
-			key = keys[i],
-			value = o[key];
+	initProps(fakeCtx, {
+		from: componentData.props,
+		store: fakeCtx,
+		saveToStore: true
+	});
 
-		Object.defineProperty(
-			fakeCtx,
-			key,
-			value !== undefined ?
-				{
-					configurable: true,
-					enumerable: true,
-					writable: true,
-					value
-				} :
+	attachMethodsFromMeta(fakeCtx);
+	implementEventAPI(fakeCtx);
+	attachAccessorsFromMeta(fakeCtx);
 
-				defProp
-		);
+	initFields(meta.systemFields, fakeCtx, fakeCtx);
+	fakeCtx.$systemFields = fakeCtx;
 
-		fakeCtx.$props[key] = value;
-	}
-
-	const defField = {
-		...defProp,
-		value: NULL
-	};
-
-	const
-		{systemFields, fields} = meta;
-
-	// Shim component fields and system fields
-	for (let list = [systemFields, fields], i = 0; i < list.length; i++) {
-		const
-			fields = list[i];
-
-		for (let keys = Object.keys(fields), i = 0; i < keys.length; i++) {
-			const
-				key = keys[i],
-				val = fields[key];
-
-			if (
-				val && (
-					val.replace !== true && (Object.isTruly(val.unique) || val.src === meta.componentName) ||
-					val.replace === false
-				)
-			) {
-				Object.defineProperty(fakeCtx, key, defField);
-			}
-		}
-	}
-
-	// Initialize values
-	initProps(fakeCtx, {store: fakeCtx, saveToStore: true});
-	initFields(systemFields, fakeCtx, fakeCtx);
-	initFields(fields, fakeCtx, fakeCtx);
-
+	initFields(meta.fields, fakeCtx, fakeCtx);
 	fakeCtx.$fields = fakeCtx;
 
 	Object.defineProperty(fakeCtx, '$$data', {
@@ -224,10 +188,12 @@ export function parseVNodeAsFlyweight(
 		}
 	});
 
-	fakeCtx.$systemFields = fakeCtx;
-	fakeCtx.hook = 'created';
+	init.createdState(fakeCtx);
 
-	const newVNode = <FlyweightVNode>execRenderObject(componentTpl.index(), fakeCtx);
+	const newVNode = <FlyweightVNode>execRenderObject(
+		componentTpl.index(),
+		fakeCtx
+	);
 
 	newVNode.fakeInstance = fakeCtx;
 	newVNode.data = newVNode.data ?? {};