+++ src/core/component/interface/component.ts
@@ -18,18 +18,6 @@ import Async, { BoundFn, ProxyCb } from 'core/async';
 
 import {
 
-	ComponentMeta,
-	Hook,
-	SyncLinkCache,
-
-	WatchPath,
-	WatchOptions,
-	RawWatchHandler
-
-} from 'core/component/interface';
-
-import {
-
 	ComponentDriver,
 
 	ComponentOptions,
@@ -40,6 +28,18 @@ import {
 	VNode
 
 } from 'core/component/engines';
+
+import {
+
+	ComponentMeta,
+	Hook,
+	SyncLinkCache,
+
+	WatchPath,
+	WatchOptions,
+	RawWatchHandler
+
+} from 'core/component/interface';
 
 /**
  * Component render function
@@ -117,9 +117,6 @@ export interface UnsafeComponentInterface<CTX extends ComponentInterface = Compo
 	$modifiedFields: CTX['$modifiedFields'];
 
 	// @ts-ignore (access)
-	$unregisteredHooks: CTX['$unregisteredHooks'];
-
-	// @ts-ignore (access)
 	$syncLinkCache: CTX['$syncLinkCache'];
 
 	// @ts-ignore (access)
@@ -157,6 +154,18 @@ export interface UnsafeComponentInterface<CTX extends ComponentInterface = Compo
 
 	// @ts-ignore (access)
 	$destroy: CTX['$destroy'];
+
+	// @ts-ignore (access)
+	onBindHook: CTX['onBindHook'];
+
+	// @ts-ignore (access)
+	onInsertedHook: CTX['onInsertedHook'];
+
+	// @ts-ignore (access)
+	onUpdateHook: CTX['onUpdateHook'];
+
+	// @ts-ignore (access)
+	onUnbindHook: CTX['onUnbindHook'];
 
 	// Internal render helpers
 
@@ -219,7 +228,17 @@ export abstract class ComponentInterface {
 	/**
 	 * Name of the active component hook
 	 */
-	readonly hook!: Hook;
+	get hook(): Hook {
+		return 'beforeRuntime';
+	}
+
+	/**
+	 * Switches the component to a new hook
+	 * @param value
+	 */
+	set hook(value: Hook) {
+		// Loopback
+	}
 
 	/**
 	 * True if the component shouldn't be destroyed after removal from DOM
@@ -391,12 +410,6 @@ export abstract class ComponentInterface {
 	protected readonly $modifiedFields!: Dictionary;
 
 	/**
-	 * Map of hook listeners that were already unregistered
-	 * (only for functional components)
-	 */
-	protected readonly $unregisteredHooks!: Dictionary<boolean>;
-
-	/**
 	 * Name of the active field to initialize
 	 */
 	protected readonly $activeField?: string;
@@ -558,5 +571,33 @@ export abstract class ComponentInterface {
 	 */
 	protected $emit(event: string, ...args: unknown[]): this {
 		return this;
+	}
+
+	/**
+	 * Handler: "bind" hook (only for functional and flyweight components)
+	 */
+	protected onBindHook(): void {
+		// Loopback
+	}
+
+	/**
+	 * Handler: "inserted" hook (only for functional and flyweight components)
+	 */
+	protected onInsertedHook(): void {
+		// Loopback
+	}
+
+	/**
+	 * Handler: "update" hook (only for functional and flyweight components)
+	 */
+	protected onUpdateHook(): void {
+		// Loopback
+	}
+
+	/**
+	 * Handler: "unbind" hook (only for functional and flyweight components)
+	 */
+	protected onUnbindHook(): void {
+		// Loopback
 	}
 }