+++ src/super/i-block/i-block.ts
@@ -48,12 +48,16 @@ import {
 
 	WatchPath,
 	RawWatchHandler,
+
+	Hook,
 	ComponentInterface,
 	UnsafeGetter,
 
 	VNode
 
 } from 'core/component';
+
+import * as init from 'core/component/construct';
 
 import 'super/i-block/directives';
 import { statuses } from 'super/i-block/const';
@@ -568,11 +572,7 @@ export default abstract class iBlock extends ComponentInterface {
 	 */
 	@computed({replace: false})
 	get componentStatus(): ComponentStatus {
-		if (this.isFlyweight) {
-			return 'ready';
-		}
-
-		return this.shadowComponentStatusStore ?? this.field.get<ComponentStatus>('componentStatusStore')!;
+		return this.shadowComponentStatusStore ?? this.field.get<ComponentStatus>('componentStatusStore') ?? 'unloaded';
 	}
 
 	/**
@@ -602,13 +602,35 @@ export default abstract class iBlock extends ComponentInterface {
 			this.field.set('componentStatusStore', value);
 		}
 
-		if (!this.isFlyweight) {
-			void this.setMod('status', value);
+		void this.setMod('component-status', value);
 
-			// @deprecated
-			this.emit(`status-${value}`, value);
-			this.emit(`componentStatus:${value}`, value, oldValue);
-			this.emit('componentStatusChange', value, oldValue);
+		// @deprecated
+		this.emit(`status-${value}`, value);
+		this.emit(`componentStatus:${value}`, value, oldValue);
+		this.emit('componentStatusChange', value, oldValue);
+	}
+
+	/** @override */
+	get hook(): Hook {
+		return this.hookStore;
+	}
+
+	/**
+	 * @override
+	 *
+	 * @param value
+	 * @emits `componentHook:{$value}(value: Hook, oldValue: Hook)`
+	 * @emits `componentHookChange(value: Hook, oldValue: Hook)`
+	 */
+	set hook(value: Hook) {
+		const
+			oldValue = this.hook;
+
+		this.hookStore = value;
+
+		if ('lfc' in this && !this.lfc.isBeforeCreate('beforeDataCreate')) {
+			this.emit(`componentHook:${value}`, value, oldValue);
+			this.emit('componentHookChange', value, oldValue);
 		}
 	}
 
@@ -794,7 +816,6 @@ export default abstract class iBlock extends ComponentInterface {
 	@system({
 		atom: true,
 		unique: true,
-		replace: true,
 		functional: false,
 		init: (ctx) => new AsyncRender(ctx)
 	})
@@ -816,7 +837,6 @@ export default abstract class iBlock extends ComponentInterface {
 	 * API to unsafe invoke of internal properties of the component.
 	 * It can be useful to create friendly classes for a component.
 	 */
-	@p({replace: true})
 	get unsafe(): UnsafeGetter<UnsafeIBlock<this>> {
 		return <any>this;
 	}
@@ -888,7 +908,7 @@ export default abstract class iBlock extends ComponentInterface {
 	 * @see [[iBlock.modsProp]]
 	 */
 	static readonly mods: ModsDecl = {
-		status: [
+		componentStatus: [
 			['unloaded'],
 			'loading',
 			'beforeReady',
@@ -918,7 +938,6 @@ export default abstract class iBlock extends ComponentInterface {
 	 */
 	@system({
 		unique: true,
-		replace: true,
 		init: (ctx) => new Daemons(ctx)
 	})
 
@@ -930,7 +949,6 @@ export default abstract class iBlock extends ComponentInterface {
 	@system({
 		atom: true,
 		unique: true,
-		replace: true,
 		init: (ctx) => new Storage(ctx)
 	})
 
@@ -942,7 +960,6 @@ export default abstract class iBlock extends ComponentInterface {
 	@system({
 		atom: true,
 		unique: true,
-		replace: true,
 		init: (ctx) => new Async(ctx)
 	})
 
@@ -955,7 +972,6 @@ export default abstract class iBlock extends ComponentInterface {
 	@system({
 		atom: true,
 		unique: true,
-		replace: true,
 		init: (ctx) => new State(ctx)
 	})
 
@@ -987,7 +1003,6 @@ export default abstract class iBlock extends ComponentInterface {
 	@system({
 		atom: true,
 		unique: true,
-		replace: true,
 		init: (ctx) => new Lazy(ctx)
 	})
 
@@ -1036,6 +1051,12 @@ export default abstract class iBlock extends ComponentInterface {
 	})
 
 	protected stageStore?: Stage;
+
+	/**
+	 * Component hook store
+	 * @see [[iBlock.hook]]
+	 */
+	protected hookStore: Hook = 'beforeRuntime';
 
 	/**
 	 * Component initialize status store
@@ -1143,7 +1164,6 @@ export default abstract class iBlock extends ComponentInterface {
 		atom: true,
 		after: 'async',
 		unique: true,
-		replace: true,
 		init: (o, d) => wrapEventEmitter(<Async>d.async, o)
 	})
 
@@ -1202,7 +1222,6 @@ export default abstract class iBlock extends ComponentInterface {
 		atom: true,
 		after: 'async',
 		unique: true,
-		replace: true,
 		init: (o, d) => wrapEventEmitter(<Async>d.async, o.r)
 	})
 
@@ -1225,7 +1244,6 @@ export default abstract class iBlock extends ComponentInterface {
 		atom: true,
 		after: 'async',
 		unique: true,
-		replace: true,
 		init: (o, d) => wrapEventEmitter(<Async>d.async, globalEmitter)
 	})
 
@@ -2078,11 +2096,6 @@ export default abstract class iBlock extends ComponentInterface {
 	@p({replace: false})
 	setMod(nodeOrName: Element | string, name: string | unknown, value?: unknown): CanPromise<boolean> {
 		if (Object.isString(nodeOrName)) {
-			if (this.isFlyweight) {
-				const ctx = this.dom.createBlockCtxFromNode(this.$el, this);
-				return Block.prototype.setMod.call(ctx, nodeOrName, name);
-			}
-
 			const res = this.lfc.execCbAfterBlockReady(() => this.block!.setMod(nodeOrName, name));
 			return res ?? false;
 		}
@@ -2111,11 +2124,6 @@ export default abstract class iBlock extends ComponentInterface {
 	@p({replace: false})
 	removeMod(nodeOrName: Element | string, name?: string | unknown, value?: unknown): CanPromise<boolean> {
 		if (Object.isString(nodeOrName)) {
-			if (this.isFlyweight) {
-				const ctx = this.dom.createBlockCtxFromNode(this.$el, this);
-				return Block.prototype.removeMod.call(ctx, nodeOrName, name);
-			}
-
 			const res = this.lfc.execCbAfterBlockReady(() => this.block!.removeMod(nodeOrName, name));
 			return res ?? false;
 		}
@@ -2357,6 +2365,7 @@ export default abstract class iBlock extends ComponentInterface {
 	 * @param ref - ref name
 	 * @param [opts] - additional options
 	 */
+	@p({replace: false})
 	protected waitRef<T = CanArray<iBlock | Element>>(ref: string, opts?: AsyncOptions): Promise<T> {
 		let
 			that = <iBlock>this;
@@ -2392,17 +2401,19 @@ export default abstract class iBlock extends ComponentInterface {
 
 		this.syncStorageState = i.syncStorageState.bind(this);
 		this.syncRouterState = i.syncRouterState.bind(this);
-
 		this.watch = i.watch.bind(this);
+
 		this.on = i.on.bind(this);
 		this.once = i.once.bind(this);
 		this.off = i.off.bind(this);
+		this.emit = i.emit.bind(this);
 	}
 
 	/**
 	 * Initializes an instance of the Block class for the current component
 	 */
 	@hook('mounted')
+	@p({replace: false})
 	protected initBlockInstance(): void {
 		if (this.block != null) {
 			const
@@ -2464,6 +2475,14 @@ export default abstract class iBlock extends ComponentInterface {
 	}
 
 	/**
+	 * Factory to create listeners of internal hook events
+	 * @param hook - hook name to listen
+	 */
+	protected createInternalHookListener(hook: string): Function {
+		return (...args) => (<Function>this[`on-${hook}-hook`.camelize(false)]).call(this, ...args);
+	}
+
+	/**
 	 * Handler: "callChild" event
 	 * @param e
 	 */
@@ -2474,6 +2493,35 @@ export default abstract class iBlock extends ComponentInterface {
 		) {
 			return e.action.call(this);
 		}
+	}
+
+	/** @override */
+	protected onBindHook(): void {
+		init.beforeMountState(this);
+
+		if (this.isFlyweight) {
+			this.componentStatus = 'ready';
+		}
+	}
+
+	/** @override */
+	protected onInsertedHook(): void {
+		init.mountedState(this);
+	}
+
+	/** @override */
+	protected onUpdateHook(): void {
+		if (this.isFlyweight) {
+			this.$el?.component?.onUnbindHook();
+		}
+
+		this.onBindHook();
+		this.onInsertedHook();
+	}
+
+	/** @override */
+	protected onUnbindHook(): void {
+		this.$destroy();
 	}
 
 	/**