+++ src/core/component/functional/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :house: Internal
+
+* Now functional components emulate hooks via the `v-hook` directive
+
 ## v3.0.0-rc.87 (2020-10-11)
 
 #### :bug: Bug Fix