+++ src/traits/i-icon/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :memo: Documentation
+
+* Improved documentation
+
 ## v3.0.0-rc.112 (2020-12-18)
 
 #### :boom: Breaking Change